+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 3.1.23
 
 * Suppress notifications from Items With Spells when updating a character.
+* Some active aura spells would not work, e.g. Cloudkill due to recursion bug with new midi arguments.
 
 # 3.1.22